siquery: move generic table querying
